requirements.txt
numpy>=1.20.0
opencv-python>=4.5.0
shapely>=1.8.0
meshpy>=2022.2.1
matplotlib>=3.5.0
scipy>=1.7.0
scikit-image>=0.19.0
vedo>=2023.4.6
Pillow>=8.0.0
vtk>=9.0.0
kneed
igl
